[BUGFIX] Replace condition for ext:hh_video_extender
The condition used to check if ext:hh_video_extender is loaded was
removed and replaced by using `ExtensionManagementUtility::isLoaded`.

Although the TypoScript setting is now added globally, this solution
seems to be more lightweight than adding a custom TypoScript condition.

This change was necessary as the condition expression could not be
parsed in case the constant was missing.